Sprint 7: redirect ingestion, worker runtime, evidence builder, E2E tooling
- Public GET /p/{slug} redirect endpoint with cookie-first visitor identity
  (IPv4 /16, IPv6 /48 coarse prefix, trusted-proxy XFF, HMAC keying)
- Production worker runtime: claim/heartbeat/dispatch/complete/fail/shutdown
- Corrected tracking sequence: baseline must precede window open;
  close_attribution_window independent of final metric scrape
- TikTokMetricsProvider protocol + Fake + PhoneAgent implementations
- EvidenceBuilder with full invariant enforcement (negative delta aborts,
  closed-window check, rich attribution_conditions JSON)
- GET /api/videos/{id}/metrics and GET /api/experiments/{id}/tracking-summary
- POST /api/videos/{id}/retry for terminal-failure recovery
- tracking_window_hours column changed to real (migration 005)
- Frontend: brief/observe/videos pages wired to API mutations, countdown
- Accelerated E2E orchestration script (scripts/accelerated_e2e.py)
- 74 unit + closeout tests pass; __pycache__ purged from source control

+++ backend/app/schemas/experiment.py
@@ -38,7 +38,7 @@ class ExperimentResponse(BaseModel):
     project_id: UUID
     hypothesis_id: UUID
     name: str
-    tracking_window_hours: int
+    tracking_window_hours: float
     status: str
     hypothesis_design_snapshot: dict
     shared_constraints: dict
